requirements.txt
apparmor==2.10.95
apparmor.click==0.3.17
apturl==0.5.2
beautifulsoup4==4.5.1
blinker==1.3
Brlapi==0.6.5
chardet==2.3.0
checkbox-support==0.22
click==0.4.45.1+16.10.20160916.0ubuntu1
command-not-found==0.3
cryptography==1.5
defer==1.0.6
distro-info===0.14build1
Django==1.11.1
django-adminlte2==0.1.6
django-datatable==0.3.1
django-registration==2.2
feedparser==5.1.3
guacamole==0.9.2
html5lib==0.999
httplib2==0.9.1
idna==2.1
Jinja2==2.8
keyring==9.1
keyrings.alt==1.1.1
language-selector==0.1
launchpadlib==1.10.4
lazr.restfulclient==0.13.4
lazr.uri==1.0.3
LibAppArmor==2.10.95
louis==2.6.5
lxml==3.6.4
Mako==1.0.4
MarkupSafe==0.23
mysqlclient==1.3.10
oauth==1.0.1
oauthlib==1.1.2
onboard==1.3.0
padme==1.1.1
pexpect==4.2.0
Pillow==3.3.1
plainbox==0.25
ptyprocess==0.5.1
pyasn1==0.1.9
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.22.0
PyJWT==1.4.1
pyparsing==2.1.8
python-apt==1.1.0b5
python-debian==0.1.29
pytz==2017.2
pyxdg==0.25
PyYAML==3.11
reportlab==3.3.0
requests==2.10.0
SecretStorage==2.3.1
sessioninstaller==0.0.0
simplejson==3.8.2
six==1.10.0
system-service==0.3
systemd-python==232
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-gdrive==0.7
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.15.1
usb-creator==0.3.3
virtualenv==15.0.3
wadllib==1.3.2
wapy==1.0.0
xdiagnose==3.8.5
xkit==0.0.0
XlsxWriter==0.7.3
zope.interface==4.2.0
